+++ run.bash
@@ -1,24 +1,24 @@
 # WT
-# cargo run --release --bin metaprofile -- \
-# -m ../methylome/within_gbM_genes/ \
-# -g ../methylome/gbM_gene_anotation_extract_Arabidopsis.bed \
-# -o /mnt/extStorage/workingDir/constantin_not_owned_by_postgres/windows/wt \
-# -e ../methylome/edgelist.txt \
-# -n ../methylome/nodelist.txt \
-# --alphabeta \
-# --name wildtype \
-# -s 1 -w 5 
+cargo run --release --bin metaprofile -- \
+-m ../methylome/within_gbM_genes/ \
+-g ../methylome/gbM_gene_anotation_extract_Arabidopsis.bed \
+-o /mnt/extStorage/workingDir/constantin_not_owned_by_postgres/windows/wt_ci \
+-e ../methylome/edgelist.txt \
+-n ../methylome/nodelist.txt \
+--alphabeta \
+--name wildtype \
+-s 1 -w 5 
 
-# # WT
-# cargo run --release --bin metaprofile -- \
-# -m ../methylome/within_gbM_genes/ \
-# -g ../methylome/gbM_gene_anotation_extract_Arabidopsis.bed \
-# -o ../windows/testing \
-# -e ../methylome/edgelist.txt \
-# -n ../methylome/nodelist.txt \
-# --alphabeta \
-# --name wildtype \
-# -s 1 -w 5 
+# # # WT
+# # cargo run --release --bin metaprofile -- \
+# # -m ../methylome/within_gbM_genes/ \
+# # -g ../methylome/gbM_gene_anotation_extract_Arabidopsis.bed \
+# # -o ../windows/testing \
+# # -e ../methylome/edgelist.txt \
+# # -n ../methylome/nodelist.txt \
+# # --alphabeta \
+# # --name wildtype \
+# # -s 1 -w 5 
 
 
 
@@ -26,23 +26,23 @@
 cargo run --release --bin metaprofile -- \
 -m /mnt/nas/zhilin/others/constantin-sergio/CMT3/total_original_methylome \
 -g ../methylome/gbM_gene_anotation_extract_Arabidopsis.bed \
--o /mnt/extStorage/workingDir/constantin_not_owned_by_postgres/windows/cmt3_new \
+-o /mnt/extStorage/workingDir/constantin_not_owned_by_postgres/windows/cmt3_ci \
 -e /home/constantin/methylome/cmt3_edgelist.txt \
 -n /home/constantin/methylome/cmt3_nodelist.txt \
 --alphabeta \
 --name cmt3 \
 -s 1 -w 5 
 
-# suv 4/5/6
-# cargo run --release --bin windows -- \
-# -m /mnt/nas/zhilin/others/constantin-sergio/SUV456/total_original_methylome \
-# -g ../methylome/gbM_gene_anotation_extract_Arabidopsis.bed \
-# -o /mnt/extStorage/workingDir/constantin_not_owned_by_postgres/windows/suv \
-# -e /mnt/nas/zhilin/others/constantin-sergio/SUV456/SUV456_edgelist.txt \
-# -n /mnt/nas/zhilin/others/constantin-sergio/SUV456/SUV456_nodelist.txt \
-# --name suv \
-# --alphabeta \
-# -s 1 -w 5 
+suv 4/5/6
+cargo run --release --bin metaprofile -- \
+-m /mnt/nas/zhilin/others/constantin-sergio/SUV456/total_original_methylome \
+-g ../methylome/gbM_gene_anotation_extract_Arabidopsis.bed \
+-o /mnt/extStorage/workingDir/constantin_not_owned_by_postgres/windows/suv_ci \
+-e /mnt/nas/zhilin/others/constantin-sergio/SUV456/SUV456_edgelist.txt \
+-n /mnt/nas/zhilin/others/constantin-sergio/SUV456/SUV456_nodelist.txt \
+--name suv \
+--alphabeta \
+-s 1 -w 5 
 
 # ros 
 #cargo run --release -- \
