+++ src/lib.rs
@@ -15,6 +15,7 @@ pub mod progress;
 pub mod setup;
 pub mod structs;
 pub mod windows;
+pub mod analysis;
 
 extern crate blas_src;
